Rename many objects

+++ TotalFitter.cs
@@ -7,8 +7,8 @@ namespace LaserMod
         public double Carrier => statisticPod.AverageValue;
         public double CarrierDispersion => statisticPod.StandardDeviation;
         public int SampleSize => (int)statisticPod.SampleSize;
-        public int MinValue => (int)statisticPod.MinimumValue;
-        public int MaxValue => (int)statisticPod.MaximumValue;
+        public double MinValue => statisticPod.MinimumValue;
+        public double MaxValue => statisticPod.MaximumValue;
         public double[] ReducedCounterData { get; }
 
         public TotalFitter(double[] counterData)
